Phase 98: close 9 critical gaps — NPN matching, CCS delay, coupling extraction, hierarchical DRC, SPICE characterizer, GNN, GDSII reader, track scalability
NPN Boolean matching with 6-input canonical form (exact <=4, heuristic 5-6),
CCS current-source delay with pi-model receiver and C_eff iteration,
spatial-grid coupling extraction, hierarchical DRC with scanline spacing,
SPICE-driven NLDM/CCS characterization with multi-corner support,
GraphSAGE GNN for congestion/timing prediction, GDSII binary stream reader,
TrackOccupancy 20-bit key (1M tracks/layer). 98/98 phases pass. ASAN clean.

+++ src/pnr/detailed_router_v2.cpp
@@ -23,7 +23,7 @@ void TrackOccupancy::clear() {
 void TrackOccupancy::insert(int layer, int track_idx, double lo, double hi, int net_id) {
     double a = std::min(lo, hi);
     double b = std::max(lo, hi);
-    int key = make_key(layer, track_idx);
+    int64_t key = make_key(layer, track_idx);
     auto& vec = tracks_[key];
     // Insert maintaining sort by lo using binary search
     Interval iv{a, b, net_id};
@@ -87,7 +87,7 @@ bool TrackOccupancy::is_free(int layer, int track_idx, double lo, double hi, int
                               double buffer, double adj_buffer, double pitch,
                               double wire_width, double spacing) const {
     // Same track check
-    int key = make_key(layer, track_idx);
+    int64_t key = make_key(layer, track_idx);
     auto it = tracks_.find(key);
     if (it != tracks_.end()) {
         if (has_overlap(it->second, lo, hi, buffer, net_id)) return false;
@@ -96,7 +96,7 @@ bool TrackOccupancy::is_free(int layer, int track_idx, double lo, double hi, int
     double perp_edge = pitch - wire_width;
     if (perp_edge < spacing) {
         for (int delta : {-1, 1}) {
-            int adj_key = make_key(layer, track_idx + delta);
+            int64_t adj_key = make_key(layer, track_idx + delta);
             auto adj_it = tracks_.find(adj_key);
             if (adj_it != tracks_.end()) {
                 if (has_overlap(adj_it->second, lo, hi, adj_buffer, net_id)) return false;
